+++ module/shell/src/main/clj/convex/shell/flow.clj
@@ -1,6 +1,11 @@
 (ns convex.shell.flow
 
+  (:import (convex.core.lang Context))
   (:require [convex.cvm :as $.cvm]))
+
+
+(set! *warn-on-reflection*
+      true)
 
 
 (declare return)
@@ -9,14 +14,20 @@
 ;;;;;;;;;;
 
 
-
 (defn fail
 
-  [ctx code message]
 
-  (return ($.cvm/exception-set ctx
-                               code
-                               message)))
+  ([^Context ctx cvm-ex]
+
+   (return (.withException ctx
+                           cvm-ex)))
+
+
+  ([ctx code message]
+
+   (return ($.cvm/exception-set ctx
+                                code
+                                message))))