+++ module/cvm/doc/changelog.md
@@ -11,6 +11,7 @@ repository.
 - Add
     - `$.cell/fake`
     - `$.cvm/exception-set`
+    - `$.cvm/exception-trace`
 
 
 ---